Draw Mandelbrot set onto image
Doesn’t actually work, but I wanted to get this out first (the module
stuff was hard to get to compile).

(We temporarily need to have an alpha channel because pixbuf.put_pixel()
assumes its existence, panicking with an out-of-bounds access otherwise.
We probably won’t use put_pixel() in the future – it does a lot of stuff
that we can pull out of the loop – so we’ll be able to remove the alpha
channel again then.)