Calculate Mandelbrot set in a separate thread
We’re operating on the lowest possible level, directly writing into the
RGB array. I think this should be safe even when we introduce actual
multi-threading – as far as I know, the CPU is supposed to handle any
problems that could arise from cache inconsistencies, as long as we’re
not writing to the same byte. (*Potentially* problematic because we’re
writing single bytes, not words. We’ll see, I guess.) The only problem
with this approach is that we never notify the pixbuf or the image of
the updates; thus, the image only updates when it has to be redrawn for
some external reason (alt+tab, popup, etc.). I assume this can easily be
fixed later.

Prior to finding this working implementation, I experienced the classic
Rust noob cycle: I fought the borrow checker (c’mon let me pass the
pixels array to the thread), lost the fight (nuh-uh must be static
lifetime), then realized there was a proper solution (thread::scoped).