Support canceling of calculation
Currently, this gets us a warning about the result of the cancel.send()
being ignored. Will be improved soon.

(Checking for a message each pixel doesn’t seem to have a significant
performance penalty. If we later find it does, or we really want to
squeeze out the last bit of performance, we can move it one loop out,
and only check once per line.)